CI: refactor neighborlist tests
