+++ test/core/neighborlist/integration/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2024-2026 Cusp AI
+# SPDX-License-Identifier: Apache-2.0
