+++ test/core/neighborlist/integration/test_against_ase.py
@@ -20,7 +20,6 @@ import pytest
 from ase.neighborlist import neighbor_list as ase_neighbor_list
 
 from kups.core.cell import Cell, PeriodicCell, TriclinicFrame, to_lower_triangular
-from kups.core.data.index import Index
 from kups.core.data.table import Table
 from kups.core.neighborlist import (
     CellListNeighborList,
@@ -28,46 +27,16 @@ from kups.core.neighborlist import (
     UniversalNeighborlistParameters,
 )
 from kups.core.result import as_result_function
-from kups.core.typing import ParticleId, SystemId
-from kups.core.utils.jax import dataclass
 
-from ..clear_cache import clear_cache  # noqa: F401
+from .._builders import EvalState, make_lh, make_systems
 
 
-@dataclass
-class _SamplePoints:
-    positions: jax.Array
-    system: Index
-    inclusion: Index
-    exclusion: Index
-
-
-@dataclass
-class _SampleSystems:
-    cell: Cell
-
-
-def _make_lh(positions: jax.Array) -> Table:
-    n = len(positions)
-    sys_keys = (0,)
-    pi_keys = tuple(ParticleId(i) for i in range(n))
-    batch = jnp.zeros(n, dtype=int)
-    return Table(
-        pi_keys,
-        _SamplePoints(
-            positions=positions,
-            system=Index(sys_keys, batch),
-            inclusion=Index(sys_keys, batch),
-            exclusion=Index.integer(jnp.arange(n, dtype=int)),
-        ),
-    )
-
-
-def _make_systems(cell: Cell, cutoff: float) -> tuple[Table, Table]:
-    sys_keys = (SystemId(0),)
-    systems = Table(sys_keys, _SampleSystems(cell=cell))
-    cutoffs = Table(sys_keys, jnp.array([float(cutoff)]))
-    return systems, cutoffs
+def _atoms_to_cell(atoms: ase.Atoms) -> Cell:
+    """Build a kups ``PeriodicCell`` from an ``ase.Atoms`` whose lattice
+    matrix is lower-triangular. Use ``_ensure_lower_triangular`` to bring
+    arbitrary ASE inputs onto this branch first."""
+    lvecs = jnp.asarray(np.asarray(atoms.cell.array))[None]
+    return PeriodicCell(TriclinicFrame.from_matrix(lvecs))
 
 
 def _ensure_lower_triangular(atoms: ase.Atoms) -> ase.Atoms:
@@ -77,7 +46,7 @@ def _ensure_lower_triangular(atoms: ase.Atoms) -> ase.Atoms:
     (a documented gradient-wrapping behaviour), so an ASE lattice that is
     not already lower-triangular would be silently truncated. Rotating up
     front via the same QR reduction kups uses internally
-    (`kups.core.cell.to_lower_triangular`) puts the input on the supported
+    (``kups.core.cell.to_lower_triangular``) puts the input on the supported
     branch and is a no-op for cells that are already lower-triangular.
     """
     cell = jnp.asarray(np.asarray(atoms.cell.array))
@@ -91,52 +60,32 @@ def _ensure_lower_triangular(atoms: ase.Atoms) -> ase.Atoms:
     return out
 
 
-def _atoms_to_cell(atoms: ase.Atoms) -> Cell:
-    """Build a kups ``PeriodicCell`` from an ASE ``Atoms`` whose lattice
-    matrix is lower-triangular. Use ``_ensure_lower_triangular`` to bring
-    arbitrary ASE inputs onto this branch first."""
-    lvecs = jnp.asarray(np.asarray(atoms.cell.array))[None]
-    return PeriodicCell(TriclinicFrame.from_matrix(lvecs))
-
-
-@dataclass
-class _EvalState:
-    """Minimal state carrying particles, systems, and neighbor-list capacity
-    hints, matching the ``IsNeighborListState`` protocol so ``from_state``
-    builds an NL with ``LensCapacity`` fields wired into ``neighborlist_params``.
-    """
-
-    particles: Table
-    systems: Table
-    neighborlist_params: UniversalNeighborlistParameters
-
-
-def _build_state(atoms: ase.Atoms, cutoff: float) -> tuple[_EvalState, Table]:
-    """Construct ``_EvalState`` from an ``ase.Atoms`` and a scalar cutoff."""
+def _build_state(atoms: ase.Atoms, cutoff: float) -> tuple[EvalState, Table]:
+    """Construct an ``EvalState`` from an ``ase.Atoms`` and a scalar cutoff."""
     positions = jnp.asarray(np.asarray(atoms.get_positions()))
-    particles = _make_lh(positions)
-    systems, cutoff_table = _make_systems(_atoms_to_cell(atoms), cutoff)
+    particles = make_lh(positions, jnp.zeros(len(atoms), dtype=int))
+    systems, cutoff_table = make_systems(
+        _atoms_to_cell(atoms), jnp.array([float(cutoff)])
+    )
     particles_per_system = Table(systems.keys, jnp.array([len(atoms)]))
     nl_params = UniversalNeighborlistParameters.estimate(
         particles_per_system, systems, cutoff_table
     )
-    state = _EvalState(
+    state = EvalState(
         particles=particles, systems=systems, neighborlist_params=nl_params
     )
     return state, cutoff_table
 
 
 def _kups_edges(nl_cls, atoms: ase.Atoms, cutoff: float):
-    """Run an NL built from a ``_EvalState`` via ``from_state``, growing its
+    """Run an NL built from an ``EvalState`` via ``from_state``, growing its
     LensCapacity-backed capacities through ``fix_or_raise`` until no assertion
-    fails. Returns ``(i, j, displacement)`` for valid (non-padding) directed
-    edges.
+    fails. Returns ``(i, j, displacement)`` for valid (non-padding) directed edges.
     """
     atoms = _ensure_lower_triangular(atoms)
     n = len(atoms)
     state, cutoff_table = _build_state(atoms, cutoff)
 
-    # Loop with a generous bound to avoid pathological growth.
     result = None
     for _ in range(2):
         nl = nl_cls.from_state(state, cutoff_table)
@@ -195,11 +144,21 @@ def _make_triclinic() -> ase.Atoms:
         positions=[[0.0, 0.0, 0.0]],
         cell=[[5.0, 0.0, 0.0], [1.5, 4.5, 0.0], [0.7, 0.4, 5.2]],
         pbc=True,
-    )  # .repeat((2,2,2))
+    )
 
 
 def _make_cu_supercell() -> ase.Atoms:
-    return ase.build.bulk("Cu", "fcc", a=3.6, cubic=True).repeat((5, 5, 5))
+    """Interior-bulk fcc Cu supercell for the minimum-image, multi-shell regime.
+
+    A 3x3x3 cubic supercell (108 atoms, box 10.8 A) is the smallest repeat that
+    reproduces the *relative* conditions that matter here, rather than raw size:
+    at cutoff 5.0 A the ratio to the half-box is 5.0/5.4 < 1 (so neighbors are
+    found by the minimum image, with no periodic image replication), the cutoff
+    spans ~3 coordination shells (2.55, 3.6, 4.41 A), and the central cell's
+    neighbors all come from adjacent cells. A 2x2x2 box (7.2 A) would fall below
+    2*cutoff and flip into the replication regime, changing the test.
+    """
+    return ase.build.bulk("Cu", "fcc", a=3.6, cubic=True).repeat((3, 3, 3))
 
 
 def _make_fcc_primitive_222() -> ase.Atoms:
@@ -215,30 +174,45 @@ def _build_bulk_al() -> ase.Atoms:
     return at
 
 
+# Cutoff sweeps are chosen by the *relative* regime they exercise, not absolute
+# size. Cost is dominated by the periodic-image replication depth,
+# ceil(cutoff / perpendicular_length): each axis emits (2*ceil+1) images, so a
+# 3-shell cutoff replicates a (multi-atom) candidate set ~343x. Because the
+# replication / stencil code (``_candidate_image_counts``,
+# ``_generate_image_offsets``) is geometry-agnostic, the deepest (3-shell)
+# regime only needs covering once — and the cheapest place is the 1-atom,
+# non-orthogonal ``triclinic`` cell (also the hardest geometry). The multi-atom
+# crystals therefore stop at their 2-shell cutoff; deeper cutoffs there were
+# pure cost, not new coverage.
 _CASES = [
-    ("cubic_Al", _build_bulk_al, (6.0, 12.0)),
-    ("cubic_Cu", _make_cubic_cu, (2.5, 5.0, 10.0)),
-    ("hcp_Mg", _make_hcp_mg, (2.5, 5.0, 10.0)),
-    ("triclinic", _make_triclinic, (3.0, 6.0, 12.0)),
-    ("Cu_5x5x5", _make_cu_supercell, (5.0,)),
-    ("fcc_primitive_222", _make_fcc_primitive_222, (3.0, 6.0)),
+    ("cubic_Al", _build_bulk_al, (6.0,)),  # rattled fcc, 2-shell replication
+    ("cubic_Cu", _make_cubic_cu, (2.5, 5.0)),  # 1- and 2-shell
+    ("hcp_Mg", _make_hcp_mg, (2.5, 5.0)),  # 1- and 2-shell
+    ("triclinic", _make_triclinic, (3.0, 6.0, 12.0)),  # deep 3-shell guard (1 atom)
+    ("Cu_3x3x3", _make_cu_supercell, (5.0,)),  # large interior-bulk, minimum-image
+    ("fcc_primitive_222", _make_fcc_primitive_222, (3.0, 6.0)),  # non-lower-triangular
+]
+
+# Flatten (system, cutoff) into one item per cutoff so the xdist scheduler can
+# distribute the expensive per-cutoff compilations across workers instead of
+# serializing a whole cutoff sweep inside a single test.
+_FLAT_CASES = [
+    (name, builder, cutoff) for name, builder, cutoffs in _CASES for cutoff in cutoffs
 ]
 
 
 class TestEnsureLowerTriangular:
     """``_ensure_lower_triangular`` is a rigid QR rotation that brings the
     ASE lattice onto the lower-triangular branch required by kups'
-    ``TriclinicFrame``. The bug it guards against is Bug #1b: without
-    pre-rotating, ``TriclinicFrame.from_matrix`` silently *projects* the
-    arbitrary basis matrix onto its lower-triangular block (a documented
-    behaviour used elsewhere for ``∂E/∂h`` gradient wrapping), which for
-    FCC primitive ASE conventions corrupts the cell into a degenerate basis
-    and triggers Bug #2 downstream.
+    ``TriclinicFrame``. Without pre-rotating, ``TriclinicFrame.from_matrix``
+    silently *projects* the arbitrary basis matrix onto its lower-triangular
+    block (a documented behaviour used elsewhere for ``∂E/∂h`` gradient
+    wrapping), which for FCC primitive ASE conventions corrupts the cell into
+    a degenerate basis.
     """
 
     def test_rotates_fcc_primitive_to_lower_triangular(self):
         atoms = ase.build.bulk("Cu", "fcc", a=3.61).repeat((2, 2, 2))
-        # Pre-condition: ASE's FCC primitive lattice is *not* lower-triangular.
         npt.assert_(
             not np.allclose(np.triu(np.asarray(atoms.cell.array), k=1), 0.0),
             "FCC primitive ASE cell is expected to be non-lower-triangular",
@@ -248,9 +222,6 @@ class TestEnsureLowerTriangular:
         npt.assert_allclose(upper, 0.0, atol=1e-5)
 
     def test_no_op_on_already_lower_triangular(self):
-        # Cubic cell is lower-triangular (diagonal). Rotation must be identity
-        # (up to optional sign flips, which keep diag positive by the QR
-        # sign-convention fix in ``to_lower_triangular``).
         atoms = ase.build.bulk("Cu", "fcc", a=3.6, cubic=True)
         rotated = _ensure_lower_triangular(atoms)
         npt.assert_allclose(
@@ -268,7 +239,6 @@ class TestEnsureLowerTriangular:
         npt.assert_allclose(v_after, v_before, rtol=1e-6)
 
     def test_preserves_pairwise_distances(self):
-        # A rigid rotation must preserve all pair distances between atoms.
         atoms = ase.build.bulk("Cu", "fcc", a=3.61).repeat((2, 2, 2))
         rotated = _ensure_lower_triangular(atoms)
         p0 = np.asarray(atoms.positions)
@@ -278,9 +248,6 @@ class TestEnsureLowerTriangular:
         npt.assert_allclose(d1, d0, atol=1e-5)
 
     def test_preserves_periodic_image_distances(self):
-        # ASE neighbor_list on the original atoms and on the rotated atoms
-        # must report the same multiset of pair distances at any cutoff
-        # (only the *direction* of each displacement vector differs).
         atoms = ase.build.bulk("Cu", "fcc", a=3.61).repeat((2, 2, 2))
         rotated = _ensure_lower_triangular(atoms)
         i0, j0, S0 = ase_neighbor_list("ijS", atoms, 6.0)
@@ -300,10 +267,6 @@ class TestEnsureLowerTriangular:
         npt.assert_allclose(np.sort(d0), np.sort(d1), atol=1e-5)
 
     def test_atoms_to_cell_produces_finite_perpendicular_lengths(self):
-        # Pre-fix, ``_atoms_to_cell`` on an FCC primitive returned a
-        # degenerate frame with ``perp = [0, NaN, NaN]``. After routing
-        # through ``_ensure_lower_triangular`` the perpendicular lengths
-        # are all finite and positive.
         atoms = ase.build.bulk("Cu", "fcc", a=3.61).repeat((2, 2, 2))
         rotated = _ensure_lower_triangular(atoms)
         cell = _atoms_to_cell(rotated)
@@ -313,9 +276,7 @@ class TestEnsureLowerTriangular:
 
 
 @pytest.mark.parametrize(
-    "case",
-    _CASES,
-    ids=[c[0] for c in _CASES],
+    "case", _FLAT_CASES, ids=[f"{n}@{c}" for n, _, c in _FLAT_CASES]
 )
 @pytest.mark.parametrize(
     "nl_cls",
@@ -324,21 +285,20 @@ class TestEnsureLowerTriangular:
 )
 class TestNeighborListAgainstASE:
     """kups neighbor lists must agree with ase.neighborlist on every (system,
-    cutoff) pair we test. Cutoff sweep covers smaller-than, larger-than, and
+    cutoff) pair we test. Cutoffs cover smaller-than, larger-than, and
     much-larger-than the perpendicular cell length."""
 
     def test_matches_ase(self, case, nl_cls):
-        name, builder, cutoffs = case
+        name, builder, cutoff = case
         atoms = builder()
-        for cutoff in cutoffs:
-            ki, kj, kdisp = _kups_edges(nl_cls, atoms, cutoff)
-            ai, aj, adisp = _ase_edges(atoms, cutoff)
+        ki, kj, kdisp = _kups_edges(nl_cls, atoms, cutoff)
+        ai, aj, adisp = _ase_edges(atoms, cutoff)
 
-            kups_canon = _canonical(ki, kj, kdisp)
-            ase_canon = _canonical(ai, aj, adisp)
+        kups_canon = _canonical(ki, kj, kdisp)
+        ase_canon = _canonical(ai, aj, adisp)
 
-            assert kups_canon == ase_canon, (
-                f"{name} @ cutoff={cutoff} {nl_cls.__name__}: "
-                f"kups produced {len(ki)} directed edges, "
-                f"ase produced {len(ai)} — displacement multisets differ"
-            )
+        assert kups_canon == ase_canon, (
+            f"{name} @ cutoff={cutoff} {nl_cls.__name__}: "
+            f"kups produced {len(ki)} directed edges, "
+            f"ase produced {len(ai)} — displacement multisets differ"
+        )
